requirements.txt
uritemplate
django
djangorestframework
markdown       
django-filter
SQLAlchemy